deb package fixes

+++ deploy/config/postrm.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
-userdel -f -r temper >/dev/null 2>&1
-deluser --group temper >/dev/null 2>&1
+if [ "$1" == "remove" ]; then
+  userdel -f -r temper >/dev/null 2>&1
+  deluser --group temper >/dev/null 2>&1
+fi
 udevadm control --reload-rules >/dev/null 2>&1
 udevadm trigger >/dev/null 2>&1
 systemctl daemon-reload >/dev/null 2>&1
